minor fixes

+++ src/main/java/ru/mail/polis/homework/simple/HomeworkTask.java
@@ -11,8 +11,7 @@ public class HomeworkTask {
      */
     public static double calcIntegral(double a, double b, ToDoubleFunction<Double> function, double delta) {
         double sum = 0;
-        while (a < b)
-        {
+        while (a < b) {
             sum += function.applyAsDouble(a) * delta;
             a += delta;
         }
@@ -25,14 +24,14 @@ public class HomeworkTask {
      */
     public static byte maxNumber(long a) {
         byte index = 0, max = 0, count = 0;
-        do{
+        do {
             if (Math.abs(a % 10) >= max) {
                 max = (byte) (a % 10);
                 index = count;
             }
             a /= 10;
             count++;
-        }while(a != 0);
+        } while (a != 0);
         return (byte) (count - index);
     }
 
@@ -44,10 +43,8 @@ public class HomeworkTask {
     public static double lineFunction(int x1, int y1, int x2, int y2, int x3) {
         if (x1 != x2) {
             return y1 - ((double) ((x1 - x3) * (y1 - y2)) / (x1 - x2));   //вывод формулы из равенста отношений (x1 - x3) / (x1 - x2) = (y1 -y3) / (y1 - y2)
-        }
-        else
-        {
-            return (y1 + y2) / 2;    //бесконечно много решений при x1 = x2 (как частное решение - точка по середине)
+        } else {
+            return (double) (y1 + y2) / 2;    //бесконечно много решений при x1 = x2 (как частное решение - точка по середине)
         }
     }
